+++ src/library/xlib/xwindows.cpp
@@ -81,14 +81,23 @@ Window XCreateWindow(Display *display, Window parent, int x, int y, unsigned int
     debuglogstdio(LCF_WINDOW, "%s call with dimensions %d x %d", __func__, width, height);
     LINK_NAMESPACE_GLOBAL(XCreateWindow);
 
+    long event_mask = 0;
+
+    /* Remove events we want to disable from the mask */
+    if (valuemask & CWEventMask) {
+        event_mask = attributes->event_mask;
+        attributes->event_mask &= ~(KeyPressMask | KeyReleaseMask);
+        attributes->event_mask &= ~(ButtonPressMask | ButtonReleaseMask | PointerMotionMask);
+    }
+
     Window w = orig::XCreateWindow(display, parent, x, y, width, height, border_width, depth, klass, visual, valuemask, attributes);
     debuglogstdio(LCF_WINDOW, "   window id is %d", w);
 
     /* Add the mask in our event queue */
     if (valuemask & CWEventMask) {
         std::shared_ptr<XlibEventQueue> queue = xlibEventQueueList.getQueue(display);
-        queue->setMask(w, attributes->event_mask);
-        debuglogstdio(LCF_WINDOW, "   event mask is %d", attributes->event_mask);
+        queue->setMask(w, event_mask);
+        debuglogstdio(LCF_WINDOW, "   event mask is %d", event_mask);
     }
 
     /* Don't save windows that has override-redirect (Wine invisible windows) */
@@ -278,6 +287,10 @@ int XSelectInput(Display *display, Window w, long event_mask)
     /* Add the mask in our event queue */
     std::shared_ptr<XlibEventQueue> queue = xlibEventQueueList.getQueue(display);
     queue->setMask(w, event_mask);
+
+    /* Remove events we want to disable from the mask */
+    event_mask &= ~(KeyPressMask | KeyReleaseMask);
+    event_mask &= ~(ButtonPressMask | ButtonReleaseMask | PointerMotionMask);
 
     return orig::XSelectInput(display, w, event_mask);
 }
@@ -501,6 +514,10 @@ int XChangeWindowAttributes(Display *display, Window w, unsigned long valuemask,
     if (valuemask & CWEventMask) {
         std::shared_ptr<XlibEventQueue> queue = xlibEventQueueList.getQueue(display);
         queue->setMask(w, attributes->event_mask);
+
+        /* Remove events we want to disable from the mask */
+        attributes->event_mask &= ~(KeyPressMask | KeyReleaseMask);
+        attributes->event_mask &= ~(ButtonPressMask | ButtonReleaseMask | PointerMotionMask);
     }
 
     return orig::XChangeWindowAttributes(display, w, valuemask, attributes);