+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Add VDPAU rendering
 * Add missing fields in Xlib input events (#365)
 * Add special case for dlsym(RTLD_NEXT) (#82)
+* Remove input events from event masks (#365)
 
 ### Changed